# Updates db indexes to improve performance