--- a/server/services/store/storetests/blocks.go
+++ b/server/services/store/storetests/blocks.go
@@ -966,12 +966,12 @@ func testGetBlockMetadata(t *testing.T, store store.Store) {
 	})
 
 	t.Run("get block history after updateAt", func(t *testing.T) {
-		rBlocks, err2 := store.GetBlocksWithType(boardID, "test")
+		rBlock, err2 := store.GetBlock("block3")
 		require.NoError(t, err2)
-		require.NotZero(t, rBlocks[2].UpdateAt)
+		require.NotZero(t, rBlock.UpdateAt)
 
 		opts := model.QueryBlockHistoryOptions{
-			AfterUpdateAt: rBlocks[2].UpdateAt,
+			AfterUpdateAt: rBlock.UpdateAt,
 			Descending:    false,
 		}
 		blocks, err = store.GetBlockHistoryDescendants(boardID, opts)