@wiggin77 2022-08-24
> getSubscriptions (MySQL 0.023 => 0.001 | PostgreSQL 14,559 ms => 1.838 ms)

Did we get any clarity on this number.  Was it really 14.6 seconds?