@mgdelacroix 2022-08-24
> > getSubscriptions (MySQL 0.023 => 0.001 | PostgreSQL 14,559 ms => 1.838 ms)
> 
> Did we get any clarity on this number. Was it really 14.6 seconds?

I think that's the Spanish locale jumping in and messing with the output of `\timings`, so I ended up using `EXPLAIN ANALYZE` to both have the output in the right format and remove the (almost negligible) network time of the query. Definitely 14ms